+++ v2rayN/ServiceLib.UdpTest/GlobalUsings.cs
@@ -1,0 +1,5 @@
+global using System.Buffers.Binary;
+global using System.Diagnostics;
+global using System.Net;
+global using System.Net.Sockets;
+global using System.Text;
