+++ v2rayN/ServiceLib.UdpTest/ServiceLib.UdpTest.csproj
@@ -1,0 +1,7 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+      <OutputType>Library</OutputType>
+  </PropertyGroup>
+
+</Project>
