+++ v2rayN/ServiceLib.UdpTest/Tester/DnsService.cs
@@ -1,6 +1,4 @@
-using System.Buffers.Binary;
-
-namespace ServiceLib.Services.Udp.Test;
+namespace ServiceLib.UdpTest.Tester;
 
 public class DnsService : IUdpTest
 {
