+++ v2rayN/v2rayN.slnx
@@ -17,6 +17,7 @@
   <Project Path="AmazTool/AmazTool.csproj" />
   <Project Path="GlobalHotKeys/src/GlobalHotKeys/GlobalHotKeys.csproj" />
   <Project Path="ServiceLib.Tests/ServiceLib.Tests.csproj" />
+  <Project Path="ServiceLib.UdpTest/ServiceLib.UdpTest.csproj" Id="77930428-4dc4-4130-9031-6b9e714f5d20" />
   <Project Path="ServiceLib/ServiceLib.csproj" />
   <Project Path="v2rayN.Desktop/v2rayN.Desktop.csproj" />
   <Project Path="v2rayN/v2rayN.csproj" />
